+++ Frameworks/Adium Framework/Source/SS_PrefsController.m
@@ -16,7 +16,7 @@
 // ************************************************
 
 
-+ (int)version
++ (NSInteger)version
 {
     return 1; // 28th June 2003
 }
